requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
semantic-version==2.9.0
    # via -r requirements.in
